distinct_sum: add hiddent tests

+++ Greedy Algorithms/Distinct Summands/src/Task.kt
@@ -1,6 +1,6 @@
 
-fun distinctSummands(n: Int): List<Int> {
-    val answer = mutableListOf<Int>()
+fun distinctSummands(n: Long): List<Long> {
+    val answer = mutableListOf<Long>()
     var left = n
     for (x in 1..n) {
         if (left - x <= x) {
